Langchain/SQLQA/requirements.txt
langchain
python-dotenv
streamlit
tiktoken
faiss-cpu
protobuf
langchain_experimental
mysql-connector-python
pymysql
sentence-transformers
chromadb==0.4.15
bitsandbytes
bitsandbytes-cuda110
accelerate